Updated API to allow multiple schema files
The original API allows only one schema file to be used. It is however
sometimes useful to have multiple schema files, for abstraction and
modularity. There was previously no way to do this in the Javascript
API. This change allows it, by specifying a list of files to be used in
the virtual file system used by empscripten.